requirements.txt
GitPython~=3.1.14